+++ src/wildboar/dimension_selection/__init__.py
@@ -1,0 +1,19 @@
+# Authors: Isak Samsten
+# License: BSD 3 clause
+"""Select a subset of dimensions"""
+
+from ._distance import (
+    SelectDimensionPercentile,
+    SelectDimensionSignificance,
+    SelectDimensionTopK,
+)
+from ._sequential import SequentialDimensionSelector
+from ._variance import DistanceVarianceThreshold
+
+__all__ = [
+    "DistanceVarianceThreshold",
+    "SequentialDimensionSelector",
+    "SelectDimensionPercentile",
+    "SelectDimensionSignificance",
+    "SelectDimensionTopK",
+]